app/requirements.txt
cassandra-driver==3.8.0
tornado==4.4.2
